+++ Client/Config/ClientConfig.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Net;
+using System.Net.Http;
 using System.Web;
 using InfluxDB3.Client.Write;
 
@@ -26,6 +27,7 @@ namespace InfluxDB3.Client.Config;
 /// <item>- Proxy: The HTTP proxy URL. Default is not set.</item>
 /// <item>- WriteOptions: Write options.</item>
 /// <item>- QueryOptions Query options.</item>
+/// <item>- HttpClient: The HttpClient will be used for Write and Query apis.</item>
 /// </list>
 ///
 /// <para>If you want create client with custom options, you can use the following code:</para>
@@ -199,6 +201,13 @@ public class ClientConfig
     /// Configuration options for query behavior in the InfluxDB client.
     /// </summary>
     public QueryOptions QueryOptions { get; set; }
+
+    /// <summary>
+    /// User-defined HttpClient.
+    /// Influxdb client will add an authentication header and base url to HttpClient. The rest is up to the users.
+    /// Users will be responsible for closing the HttpClient.
+    /// </summary>
+    public HttpClient? HttpClient { get; set; }
 
     internal void Validate()
     {