+++ Client/InfluxDBClient.cs
@@ -343,7 +343,7 @@ namespace InfluxDB3.Client
             config.Validate();
 
             _config = config;
-            _httpClient = CreateAndConfigureHttpClient(_config);
+            _httpClient = CreateOrGetHttpClient(_config);
             FlightSqlClient = new FlightSqlClient(config: _config, httpClient: _httpClient);
             _restClient = new RestClient(config: _config, httpClient: _httpClient);
             _gzipHandler = new GzipHandler(config.WriteOptions != null ? config.WriteOptions.GzipThreshold : 0);
@@ -840,7 +840,12 @@ namespace InfluxDB3.Client
 
         public void Dispose()
         {
-            _httpClient.Dispose();
+            // _config.HttpClient == null means HttpClient is created by the library, not from the user.
+            // so the client will be responsible for disposing of the HttpClient.
+            if (_config.HttpClient == null)
+            {
+                _httpClient.Dispose();
+            }
             FlightSqlClient.Dispose();
             _disposed = true;
         }
@@ -876,7 +881,28 @@ namespace InfluxDB3.Client
             return sb;
         }
 
-        internal static HttpClient CreateAndConfigureHttpClient(ClientConfig config)
+        internal static HttpClient CreateOrGetHttpClient(ClientConfig config)
+        {
+            var httpClient = config.HttpClient;
+            if (httpClient == null)
+            {
+                httpClient = CreateHttpClient(config);
+            }
+
+            if (httpClient.BaseAddress == null)
+            {
+                httpClient.BaseAddress = new Uri(config.Host);
+            }
+
+            if (!string.IsNullOrEmpty(config.Token))
+            {
+                _setAuthenticationHeader(httpClient, config);
+            }
+
+            return httpClient;
+        }
+
+        private static HttpClient CreateHttpClient(ClientConfig config)
         {
             var handler = new HttpClientHandler();
             if (handler.SupportsRedirectConfiguration)
@@ -916,15 +942,15 @@ namespace InfluxDB3.Client
             {
                 Timeout = config.Timeout
             };
-
             client.DefaultRequestHeaders.UserAgent.ParseAdd(AssemblyHelper.GetUserAgent());
-            if (!string.IsNullOrEmpty(config.Token))
-            {
-                string authScheme = string.IsNullOrEmpty(config.AuthScheme) ? "Token" : config.AuthScheme!;
-                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(authScheme, config.Token);
-            }
 
             return client;
+        }
+
+        private static void _setAuthenticationHeader(HttpClient httpClient, ClientConfig config)
+        {
+            var authScheme = string.IsNullOrEmpty(config.AuthScheme) ? "Token" : config.AuthScheme!;
+            httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(authScheme, config.Token);
         }
 
         private static string OptionMessage(string property)